fix(mutants): the gate could not fail, and could not see its own pure logic
`Mutation Testing (incremental)` was red at 52% with 220 survivors. Three
separate defects, one of them new.

**The Display/Debug filter matched nothing.** `mutants.toml` carried
`exclude_re = ["^fmt::"]`, commented "Display/Debug impls". cargo-mutants
matches that regex against the whole line it prints for a mutant —
`src/x.rs:1: replace <impl core::fmt::Debug for T>::fmt -> … with …` — which
begins with the file path, so a pattern anchored at `fmt::` can never match.
Seventeen unkillable `Debug::fmt` mutants had survived every sweep. The form
cargo-mutants documents, `impl Debug`, does not match this crate either: it
writes `impl core::fmt::Debug for T`, and the qualified path lands in the name.
The pattern is now `impl [a-z:]*Debug for`. `Display` is deliberately not
excluded — those impls render error text that tests assert on, so their mutants
die and belong in the gate.

**The incremental job did not re-apply `exclude_re`.** It already re-applies
`exclude_globs` as CLI flags, because `--file` bypasses the config; cargo-mutants
only *combines* a CLI `--exclude-re` with the config file from 27.0.0 onwards,
and this job passes one. Both are re-applied now, keeping mutants.toml
authoritative either side of that change.

**`value.rs` and `query.rs` accounted for 206 of the 220.** Every function left
in them wraps a DuckDB C call, which a `--cargo-arg=--lib` run with no live
engine structurally cannot reach — the rationale mutants.toml already carried
for nine sibling modules. Excluding them wholesale would have swallowed their
pure code too, so it moved to siblings that stay in the gate:

  src/value/hugeint.rs   the four 128-bit word helpers
  src/value/defaults.rs  the fourteen as_*_or accessors
  src/query/cstr.rs      to_c_sql, c_str_to_owned

Seven of the `as_*_or` accessors had no unit test at all, and nothing exercised
`c_str_to_owned`'s non-null or non-UTF-8 paths — all of them were among the 220.
Nine new tests turn those survivors into kills instead of hiding them. No public
item moved; the re-export keeps every existing path.

Two survivors were worth more than an annotation:

- `SecretEntry`'s zeroize-on-drop is a security property the crate advertises in
  its module docs, and nothing asserted it. The body is now
  `SecretEntry::zeroize_in_place`, tested directly, with `Drop` a one-line
  delegation.
- `hugeint_to_i128` and `uhugeint_to_u128` combined their halves with `|`.
  The halves occupy disjoint bits, so `|` -> `^` cannot change the result for
  any input: an equivalent mutant, unkillable by construction. They add instead.
  `+` is identical here, cannot overflow (`i64::MIN << 64` is exactly
  `i128::MIN`; `u64::MAX` in both halves is exactly `u128::MAX`, and the round
  trips at those extremes would panic in a debug build if that reasoning were
  wrong), and `-` or `*` in its place dies at once.

The rest are annotated in the source with `#[mutants::skip]` and a reason:
the bare FFI reads (`DataChunk::size` / `column_count`, `ScalarBindData::set`,
`ScalarLocalState::set`), two `Drop` impls whose effect is visible only in freed
memory or to a leak checker (`OwnedVector`, `SecretEntry`), the deprecated
`FfiBindData::get_from_bind` whose `with None` mutant *is* the function, `map2`
and `map2_str` whose per-row NULL check only runs inside DuckDB's expression
executor, and `Value::as_str_or_default`, whose null-handle answer is exactly
the mutant's `String::new()`.

Verified locally with the job's own invocation over the 40 changed source
files: 404 mutants, 250 caught, 154 unviable, none missed.
